Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
praw==7.7.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
channels==4.0.0 